# Improve `BindableList` performance + reduce allocations